lib/count_transactions_depository/requirements.txt
pngme-api == 0.10.0
